Bump `windows(-core)` crate range to `0.57-0.59`
And regenerate with `windows-bindgen 0.59`.  This emits some
significantly newer code, and uses a new `Ref` type that's only
available since `windows 0.57`.  Also remove the `"implement"` feature
which we never relied on on and which appears to be gone now in `windows
0.59`.
