+++ Cargo.toml
@@ -25,11 +25,12 @@ all-features = true
 [dependencies]
 libloading = { version = "0.8", optional = true }
 once_cell = { version = "1.19", optional = true }
-windows = { version = ">=0.56, <=0.58", features = ["implement", "Win32_Foundation", "Win32_Graphics_Direct3D12", "Win32_Storage_FileSystem"] }
-windows-core = { version = ">=0.56, <=0.58" }
+windows = { version = ">=0.57, <=0.59", features = ["Win32_Foundation", "Win32_Graphics_Direct3D12", "Win32_Storage_FileSystem"], default-features = false }
+windows-core = ">=0.57, <=0.59"
+windows-targets = "0.53"
 
 [dev-dependencies]
-windows = { version = ">=0.56, <=0.58", features = ["Win32_Foundation", "Win32_Graphics_Direct3D12", "Win32_Graphics_Direct3D", "Win32_Graphics_Dxgi_Common", "Win32_System_WindowsProgramming", "Win32_Security", "Win32_Storage_FileSystem", "Win32_System_Com", "Win32_System_Threading"] }
+windows = { version = ">=0.57, <=0.59", features = ["Win32_Foundation", "Win32_Graphics_Direct3D12", "Win32_Graphics_Direct3D", "Win32_Graphics_Dxgi_Common", "Win32_System_WindowsProgramming", "Win32_Security", "Win32_Storage_FileSystem", "Win32_System_Com", "Win32_System_Threading"], default-features = false }
 
 [workspace]
 members = [
