+++ api_gen/Cargo.toml
@@ -5,4 +5,4 @@ edition = "2021"
 publish = false
 
 [dependencies]
-windows-bindgen = "0.58"
+windows-bindgen = "0.59"
