+++ api_gen/src/main.rs
@@ -1,7 +1,5 @@
-use windows_bindgen::{bindgen, Result};
+use windows_bindgen::bindgen;
 
-fn main() -> Result<()> {
-    let log = bindgen(["--etc", "bindings.txt"])?;
-    println!("{}", log);
-    Ok(())
+fn main() {
+    bindgen(["--etc", "bindings.txt"])
 }
